Handle case where verification script requested no changes
In the current state, the verification script will exit with an error if
the requested database value is not different from the old value. This
is not really an error condition and should not result in the
verification failing. Instead, the number of stars on the returned web
page are counted to verify that the correct number is returned.